# docs/reservation-openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Events
  description: Operations with events
- name: Reservations
  description: Operations with reservations
paths:
  /api/v1/reserve/{id}:
    put:
      tags:
      - Reservations
      operationId: updateReservation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReservationRequestDto'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationDto'
  /api/v1/event/{id}:
    get:
      tags:
      - Events
      operationId: getEventById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventDto'
    put:
      tags:
      - Events
      operationId: updateEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventRequestDto'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventDto'
  /api/v1/reserve:
    get:
      tags:
      - Reservations
      operationId: getUserReserves
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationWithEventDto'
    post:
      tags:
      - Reservations
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequestDto'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationCreatedDto'
  /api/v1/event:
    get:
      tags:
      - Events
      operationId: getAllEvents
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventDto'
    post:
      tags:
      - Events
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequestDto'
        required: true
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EventCreatedDto'
  /api/v1/reserve/{reservationId}:
    get:
      tags:
      - Reservations
      operationId: getReservationById
      parameters:
      - name: reservationId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReservationWithEventDto'
  /api/v1/event/{id}/reserve:
    get:
      tags:
      - Events
      operationId: getEventReserves
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                oneOf:
                - $ref: '#/components/schemas/RestApiErrorDto'
                - $ref: '#/components/schemas/RestApiValidationErrorDto'
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RestApiErrorDto'
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
components:
  schemas:
    RestApiValidationErrorDto:
      type: object
      properties:
        error:
          type: string
        violations:
          type: object
          additionalProperties:
            type: string
    RestApiErrorDto:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
    UpdateReservationRequestDto:
      required:
      - isCanceled
      - seats
      type: object
      properties:
        seats:
          minimum: 1
          type: integer
          format: int32
        isCanceled:
          type: boolean
    ReservationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        eventId:
          type: string
          format: uuid
        seats:
          type: integer
          format: int32
        isCanceled:
          type: boolean
    UpdateEventRequestDto:
      required:
      - durationSeconds
      - isCanceled
      - name
      - startTime
      - ticketPrice
      - totalSeats
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 0
          type: string
        description:
          maxLength: 2000
          minLength: 0
          type: string
        startTime:
          type: string
          format: date-time
        durationSeconds:
          minimum: 1
          type: integer
          format: int64
        ticketPrice:
          minimum: 0
          type: number
          format: double
        totalSeats:
          minimum: 1
          type: integer
          format: int32
        isCanceled:
          type: boolean
    EventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          format: int64
        ticketPrice:
          type: number
          format: double
        totalSeats:
          type: integer
          format: int32
        isCanceled:
          type: boolean
    CreateReservationRequestDto:
      required:
      - eventId
      - seats
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        seats:
          minimum: 1
          type: integer
          format: int32
    ReservationCreatedDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
    CreateEventRequestDto:
      required:
      - durationSeconds
      - name
      - startTime
      - ticketPrice
      - totalSeats
      type: object
      properties:
        name:
          maxLength: 255
          minLength: 0
          type: string
        description:
          maxLength: 2000
          minLength: 0
          type: string
        startTime:
          type: string
          format: date-time
        durationSeconds:
          minimum: 1
          type: integer
          format: int64
        ticketPrice:
          minimum: 0
          type: number
          format: double
        totalSeats:
          minimum: 1
          type: integer
          format: int32
    EventCreatedDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
    ReservationEventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        durationSeconds:
          type: integer
          format: int64
        ticketPrice:
          type: number
          format: double
        isCanceled:
          type: boolean
    ReservationWithEventDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        event:
          $ref: '#/components/schemas/ReservationEventDto'
        seats:
          type: integer
          format: int32
        isCanceled:
          type: boolean